+++ Backend/sprest-mongo/src/main/java/ofm/FileRepository.java
@@ -6,10 +6,9 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FileRepository extends MongoRepository < FileModel, Long > {
-    
+public interface FileRepository extends MongoRepository < FileModel, Integer > {
 	// return a list of files that has same belFolderId
-	List<FileModel> findByBelFolderId(long belFolderId);
+	List<FileModel> findByBelFolderId(Integer belFolderId);
     
 	// return a list of files that has same file name
 	List<FileModel> findByName(String name);
